Refatorado o fluxo do caso de uso que registra um novo Client
