+++ ms-clientes/src/main/java/io/github/paulovieirajr/msclientes/core/application/usecases/UseCase.java
@@ -1,6 +1,6 @@
 package io.github.paulovieirajr.msclientes.core.application.usecases;
 
-public interface ClientUseCaseHandler <T, R> {
+public interface UseCase<T, R> {
 
     R execute(T t);
 }
