+++ ms-clientes/src/main/java/io/github/paulovieirajr/msclientes/infra/config/ClientBeansConfig.java
@@ -2,6 +2,7 @@ package io.github.paulovieirajr.msclientes.infra.config;
 
 import io.github.paulovieirajr.msclientes.core.application.repositories.ClientRepositoryGateway;
 import io.github.paulovieirajr.msclientes.core.application.usecases.ClientRegisterUseCase;
+import io.github.paulovieirajr.msclientes.core.application.usecases.ClientRegisterUseCaseHandler;
 import io.github.paulovieirajr.msclientes.infra.gateways.ClientRepositoryGatewayJpa;
 import io.github.paulovieirajr.msclientes.infra.mapper.ClientMapper;
 import io.github.paulovieirajr.msclientes.infra.repository.ClientRepository;
@@ -9,11 +10,11 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
-public class BeansConfig {
+public class ClientBeansConfig {
 
     @Bean
     public ClientRegisterUseCase clientRegisterUseCase(ClientRepositoryGateway clientRepositoryGateway) {
-        return new ClientRegisterUseCase(clientRepositoryGateway);
+        return new ClientRegisterUseCaseHandler(clientRepositoryGateway);
     }
 
     @Bean
