+++ ms-clientes/src/test/java/io/github/paulovieirajr/msclientes/api/controller/ClientRegisterControllerTest.java
@@ -2,7 +2,7 @@ package io.github.paulovieirajr.msclientes.api.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.github.paulovieirajr.msclientes.api.dto.ClientRequest;
-import io.github.paulovieirajr.msclientes.core.application.usecases.ClientRegisterUseCase;
+import io.github.paulovieirajr.msclientes.core.application.usecases.ClientRegisterUseCaseHandler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -21,7 +21,7 @@ class ClientRegisterControllerTest {
     MockMvc mockMvc;
 
     @MockBean
-    ClientRegisterUseCase clientRegisterUseCase;
+    ClientRegisterUseCaseHandler clientClientRegisterUseCaseHandler;
 
     @Nested
     @DisplayName("Create Client with correct values")
