+++ ms-clientes/src/test/java/io/github/paulovieirajr/msclientes/core/usecases/ClientClientRegisterUseCaseHandlerTest.java
@@ -2,7 +2,7 @@ package io.github.paulovieirajr.msclientes.core.usecases;
 
 
 import io.github.paulovieirajr.msclientes.api.dto.ClientRequest;
-import io.github.paulovieirajr.msclientes.core.application.usecases.ClientRegisterUseCase;
+import io.github.paulovieirajr.msclientes.core.application.usecases.ClientRegisterUseCaseHandler;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -14,16 +14,16 @@ import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith(SpringExtension.class)
 @SpringBootTest
-class ClientRegisterUseCaseTest {
+class ClientClientRegisterUseCaseHandlerTest {
 
     @Autowired
-    ClientRegisterUseCase clientRegisterUseCase;
+    ClientRegisterUseCaseHandler clientClientRegisterUseCaseHandler;
 
     @Test
     @DisplayName("Should return a a client response when client is created")
     void createClient() {
         var clientRequest = new ClientRequest("John Doe", "56010962013", 30);
-        var clientResponse = clientRegisterUseCase.execute(clientRequest);
+        var clientResponse = clientClientRegisterUseCaseHandler.execute(clientRequest);
         assertNotNull(clientResponse);
     }
 }
